requirements.txt
python-dotenv
openai
langchain
tiktoken
pandas
docarray
duckduckgo-search
wikipedia
DateTime